+++ testdata/xlb/goldens/tsout/es-419.ts
@@ -1,11 +1,17 @@
 
     // Do not modify this file by hand!
     // Re-generate this file by running lit-localize
+
     import {html} from 'lit-html';
 
     /* eslint-disable no-irregular-whitespace */
+    /* eslint-disable @typescript-eslint/camelcase */
+    /* eslint-disable @typescript-eslint/no-explicit-any */
+
     export const messages = {
-      a: `Hola Galaxia!`,
-b: html`Hola <b>Mundo!</b>`,
+      lit: html`Hola <b>Mundo!</b>`,
+lit_variables: (url: any, name: any) => html`Hola ${name}, clic <a href="${url}">aquí</a>!`,
+string: `Hola Mundo!`,
+variable: (name: any) => `Hola ${name}!`,
     };