+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Add support for variables:
+
+  ```typescript
+  msg(
+    'hello',
+    (url: string, name: string) =>
+      html`Hello ${name}, click <a href="${url}">here</a>!`,
+    'World',
+    'https://www.example.com/'
+  );
+  ```
 
 ## 0.1.1